pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "epopy"
version = "0.1.0"
description = "Async Python client for EPO OPS API"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.12"
authors = [
    { name = "Quik", email = "quinten@daskju.com" }
]
keywords = ["epo", "ops", "patent", "api", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "xmltodict>=0.13.0",
    "lxml>=5.0.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "respx>=0.20.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.0.0",
    "types-lxml>=2026.1.1",
]

[project.urls]
Homepage = "https://github.com/tu-po/epopy"
Repository = "https://github.com/tu-po/epopy"
Documentation = "https://github.com/tu-po/epopy#readme"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/epopy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
strict = true
ignore_missing_imports = true
